Dashboard Track stat OK, different dashboard in portret and landscape mode

+++ trunk/RMaps.AppLib/src/com/robert/maps/applib/overlays/CurrentTrackOverlay.java
@@ -218,10 +218,11 @@ public class CurrentTrackOverlay extends TileViewOverlay {
         }
 
 		@Override
-		public void onTrackStatUpdate(int Cnt, double Distance, double Duration, double MaxSpeed, double AvgSpeed,
-				int MoveTime, double AvgMoveSpeed) throws RemoteException {
+		public void onTrackStatUpdate(int Cnt, double Distance, long Duration, double MaxSpeed, double AvgSpeed, long MoveTime, double AvgMoveSpeed)
+				throws RemoteException {
 			
 		}
+
     };
 
 }
